[#96] utils.js로 유틸함수 옮기기
