+++ src/components/FeedCardAnswerEdit.jsx
@@ -1,7 +1,7 @@
-import { useState, useEffect } from 'react'
+import { useEffect, useState } from 'react'
 import { FillButton } from './Buttons'
 import { Textarea } from './Inputs'
-import getRelativeDate from '../utils/getRelativeDate'
+import { getRelativeDate } from '../modules/utils'
 import styles from './FeedCardAnswerEdit.module.css'
 import postNewAnswer from '../api/postNewAnswer'
 import editAnswer from '../api/editAnswer'
